Move Pretty to Options
This is one more step towards fixing orphan instances.
